chore: 更新项目结构，添加新的仪表盘布局和接口管理功能，重构样式和组件，优化依赖管理，增强用户体验
