+++ clients/atom/app/providers.tsx
@@ -1,0 +1,31 @@
+'use client';
+
+import React from 'react';
+import { createCache, extractStyle, StyleProvider } from '@ant-design/cssinjs';
+import { ConfigProvider } from 'antd';
+import { useServerInsertedHTML } from 'next/navigation';
+import zhCN from 'antd/locale/zh_CN';
+
+export function AntdRegistry({ children }: { children: React.ReactNode }) {
+  const cache = React.useMemo(() => createCache(), []);
+  const isServerInserted = React.useRef<boolean>(false);
+
+  useServerInsertedHTML(() => {
+    if (isServerInserted.current) {
+      return;
+    }
+    isServerInserted.current = true;
+    return (
+      <style
+        id="antd"
+        dangerouslySetInnerHTML={{ __html: extractStyle(cache, true) }}
+      />
+    );
+  });
+
+  return (
+    <StyleProvider cache={cache}>
+      <ConfigProvider locale={zhCN}>{children}</ConfigProvider>
+    </StyleProvider>
+  );
+}
