+++ clients/atom/package.json
@@ -9,19 +9,24 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@ant-design/cssinjs": "^1.19.0",
     "@ant-design/icons": "^5.3.0",
     "@ant-design/pro-components": "^2.6.49",
     "@monaco-editor/react": "^4.6.1",
     "antd": "^5.15.0",
+    "axios": "^1.6.7",
+    "dayjs": "^1.11.10",
+    "lodash": "^4.17.21",
     "monaco-editor": "^0.47.0",
     "next": "^14.2.1",
     "qs": "^6.11.2",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-monaco-editor": "^0.55.0"
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
+    "@types/lodash": "^4.14.202",
     "@types/node": "^20.12.5",
+    "@types/qs": "^6.9.12",
     "@types/react": "^18.2.58",
     "@types/react-dom": "^18.2.19",
     "@typescript-eslint/eslint-plugin": "^7.1.0",
