+++ clients/atom/public/images/cli-logo.png
