+++ clients/atom/public/images/component-logo.png
