+++ clients/atom/public/images/design-logo.png
