+++ clients/atom/public/images/devops-logo.png
